Pseudocode as code
I spent a while this evening hammering away at the commands we'll need
to make this puppy work, and implemented a bunch of stuff… with the big
caveat that my numbers are almost entirely all made up.

LOOK IN RobotConstants.java!

ADJUST NUMBERS IN RobotConstants.java!

I didn't tackle the winch, since I have some questions about how we
know when to stop winding.

I did, however, spec out the El Toro commands, as well as the basic
overview of firing the catapult and autonomous (highly dependent on
syncing up encoders and vision to really work.)