Winch Logic
After discussion with Kevin, we think that we can set the motor speed
once in the initialize() method and forget about it (which means that
the commands to start/stop the El Toro arm twirling don’t need to be
running indefinitely — they just need to run quickly).

Updated winch and dogbox code on the assumption that we will have a
limit switch at the bottom of the catapult and a quadrature encoder on
the winch engage window motor.