El Toro to Flipper Conversions
From David:

Went into RoboBuilder and made a new subsystem for the flipper. Moved
the "eltoro" updown motor into the flipper, where it now belongs, and
put proper limit switches in.

Added the final limit switch of the bot into the code using RoboBuilder
and renamed the limit switches to make sense.

In the code: Deleted any classes of and relating to the raising and
lowering of eltoro. Replaced them with Raising and Lowering of the
Flipper, and made a new EjectBall class. No code has been written for
any of the new classes, and that does need to be done.