Simplify! Subsystematize!
Stripped out a bunch of extraneous commands that were only needed so
that we could have a big, fancy command group that did what we really
wanted. Converted those commands into subsystem methods that could be
called by a (much simpler) command that did what we wanted.

Reimplemented Flipper and El Toro subsystems to use this approach. Note
the lack of references to RobotMap -- we're actually now using the
subsystems as they were meant to be used!