Vomit Ball
Rescripted based on current position of El Toro. Get the arms behind
the ball (while spinning such that they will force the ball away from
the catapult as they move towards the catapult themselves). Then push
out with the arms to the park position, still spinning, to “vomit” the
ball off of the front of the catapult.