Removed Potential Blocking
Rather than using a potentially blocking while loop in the Autonomous
code, trying out an subsystems based idea.

Essentially, Pause for distance is a blocking command that won’t let
the next thing happen until the drive train (presumed to be in motion
already) hits a particular distance. So we queue that up before
FireCatapult, and we should be able to let the subsystem loops sort it
all out.