Track engaged/disengaged
Added some code that should, theoretically, track whether or not the
dogbox is engaged or disengaged and prevent it from accidentally forcing
itself past its limit.

CAVEAT EMPTOR: This starts with the assumption that the dogbox always
powers on engaged (as in autonomous). This may require some fiddling if
it is not true!