Safer Catapult Wind
This is one approach to avoiding El Toro encirclement by the catapult: block raising El Toro if the catapult is currently winding.

This approach should be 100% safe, but there are other, dicier approaches, such as using timeouts to only block raising El Toro for part of the wind — or to only raise El Toro partway while winding is happening. Those approaches would require some timing of processes to get good numbers for the timeouts.

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>