Variable Autonomous
One or two ball autonomous code (choose which in RobotConstants).

Cleaned up a few unused commands as well.

CAVEAT EMPTOR: If using the RobotBuilder, beware that that Robot.java
must be modified to pass RobotConstants.AUTONOMOUS_BALLS as a parameter
to utonomous(). Otherwise, the autonomous mode will revert to the
default single ball mode.